test: merge from_yaml/to_yaml tests into test_yaml_roundtrip
test_from_yaml_errors.py (schema-drift hints, dict-input hardening) and
test_to_yaml_roundtrip.py (lossless round-trips, clean exports,
serializer-warning guards) shared fixtures and covered the two
directions of the same SUEWSConfig YAML I/O surface. Bodies unchanged;
23 collected items before and after.
