+++ test/data_model/test_yaml_roundtrip.py
@@ -1,7 +1,16 @@
-"""
-Round-trip tests for SUEWSConfig.to_yaml() output.
+"""SUEWSConfig YAML I/O round-trips and error handling.
 
-Verifies both lossless round-tripping and clean-export behaviour.
+One file for the from_yaml/to_yaml surface:
+
+* ``from_yaml`` error handling (formerly ``test_from_yaml_errors.py``):
+  schema-drift hints added in #1303 (TypeError wrapping, extra_forbidden
+  hints, detected-vs-current version naming) and the gh#1530 dict-input
+  hardening follow-ups (safe YAML loading, loud unknown keys, validated
+  assignment).
+
+* ``to_yaml`` round-trips (formerly ``test_to_yaml_roundtrip.py``):
+  lossless round-tripping, clean exports, and the gh#1569 guarantee that
+  serialising a config emits no Pydantic serializer warnings.
 """
 
 import tempfile
@@ -10,13 +19,215 @@ from pathlib import Path
 
 import pytest
 import yaml
+
 from supy._env import trv_supy_module
 from supy.data_model.core.config import SUEWSConfig
 from supy.data_model.core.type import RefValue
+from supy.data_model.configuration import CURRENT_SCHEMA_VERSION
 from supy.data_model.validation.pipeline.phase_a import find_extra_parameters
 
 pytestmark = pytest.mark.api
 
+
+# ===========================================================================
+# from_yaml error handling -- formerly test_from_yaml_errors.py
+# ===========================================================================
+
+@pytest.fixture
+def sample_config_dict() -> dict:
+    """Load the packaged sample configuration as a mutable dict."""
+    path = trv_supy_module / "sample_data" / "sample_config.yml"
+    with path.open(encoding="utf-8") as f:
+        return yaml.safe_load(f)
+
+
+def _write_yaml(tmp_path: Path, data: dict) -> Path:
+    """Serialise `data` to a YAML file under `tmp_path` and return the path."""
+    target = tmp_path / "drifted.yml"
+    with target.open("w", encoding="utf-8") as f:
+        yaml.safe_dump(data, f, sort_keys=False)
+    return target
+
+
+@pytest.mark.cfg
+class TestFromYamlDriftHints:
+    """Parse failures surface actionable schema-drift hints."""
+
+    def test_valid_yaml_still_parses(self, sample_config_dict, tmp_path):
+        """Positive control: the packaged sample still loads cleanly."""
+        # ARRANGE
+        path = _write_yaml(tmp_path, sample_config_dict)
+
+        # ACT
+        config = SUEWSConfig.from_yaml(str(path))
+
+        # ASSERT
+        assert config.name == "sample_config"
+
+    def test_extra_forbidden_gets_drift_hint(self, sample_config_dict, tmp_path):
+        """A stray field on a nested `extra=forbid` model triggers the hint.
+
+        `SUEWSConfig` itself uses `extra="allow"`, but nested models
+        (e.g. `SiteProperties`) forbid extras, so we trigger the drift smell
+        there.
+        """
+        # ARRANGE
+        drifted = dict(sample_config_dict)
+        drifted["sites"][0]["properties"]["removed_field_from_old_release"] = 42
+        path = _write_yaml(tmp_path, drifted)
+
+        # ACT / ASSERT
+        with pytest.raises(ValueError) as excinfo:
+            SUEWSConfig.from_yaml(str(path))
+
+        message = str(excinfo.value)
+        assert "suews-convert -i <old.yml>" in message
+        assert f"Current schema version:  {CURRENT_SCHEMA_VERSION}" in message
+        assert "Detected schema version:" in message
+
+    def test_typeerror_path_gets_drift_hint(self, sample_config_dict, tmp_path):
+        """A raw TypeError from validation is wrapped with the drift hint."""
+        # ARRANGE: force a TypeError through from_yaml. We monkeypatch the
+        # config-construction step to mimic what union dispatch does when
+        # drifted YAML reaches a model with a kwarg-strict __init__.
+        path = _write_yaml(tmp_path, sample_config_dict)
+
+        original_init = SUEWSConfig.__init__
+
+        def raise_type_error(self, **_):
+            raise TypeError(
+                "RefValue.__init__() got an unexpected keyword argument 'working_day'"
+            )
+
+        SUEWSConfig.__init__ = raise_type_error
+        try:
+            with pytest.raises(ValueError) as excinfo:
+                SUEWSConfig.from_yaml(str(path))
+        finally:
+            SUEWSConfig.__init__ = original_init
+
+        # ASSERT
+        message = str(excinfo.value)
+        assert "suspected schema drift" in message
+        assert "suews-convert -i <old.yml>" in message
+        assert "RefValue.__init__()" in message
+        assert f"Current schema version:  {CURRENT_SCHEMA_VERSION}" in message
+
+    def test_drift_hint_names_detected_schema_version(
+        self, sample_config_dict, tmp_path
+    ):
+        """When the YAML declares schema_version, it shows up as 'Detected'."""
+        # ARRANGE
+        drifted = dict(sample_config_dict)
+        drifted["schema_version"] = "2025.12"
+        drifted["sites"][0]["properties"]["from_an_old_release"] = True
+        path = _write_yaml(tmp_path, drifted)
+
+        # ACT / ASSERT
+        with pytest.raises(ValueError) as excinfo:
+            SUEWSConfig.from_yaml(str(path))
+
+        message = str(excinfo.value)
+        assert "Detected schema version: 2025.12" in message
+
+    def test_unsigned_yaml_hint_requests_from_ver(self, sample_config_dict, tmp_path):
+        """Unsigned YAMLs must not be told to run the bare suews-convert command.
+
+        The CLI rejects unsigned YAMLs unless `-f/--from` is supplied, so
+        the hint must advertise that flag and must not claim a spurious
+        "Detected schema version: <current>" that was only there because
+        from_yaml stamped CURRENT_SCHEMA_VERSION as a default.
+        """
+        # ARRANGE - strip the signature to mimic a pre-schema-version release
+        drifted = dict(sample_config_dict)
+        drifted.pop("schema_version", None)
+        drifted["sites"][0]["properties"]["from_an_old_release"] = True
+        path = _write_yaml(tmp_path, drifted)
+
+        # ACT / ASSERT
+        with pytest.raises(ValueError) as excinfo:
+            SUEWSConfig.from_yaml(str(path))
+
+        message = str(excinfo.value)
+        assert "No schema_version field in YAML" in message
+        assert "-f <release-tag>" in message
+        assert f"Detected schema version: {CURRENT_SCHEMA_VERSION}" not in message
+
+
+class TestDictInputHardening:
+    """gh#1530 follow-ups: safe YAML loading and loud unknown-key handling."""
+
+    def test_from_yaml_rejects_python_object_tags(self, sample_config_dict, tmp_path):
+        """from_yaml must not construct arbitrary Python objects from tags.
+
+        yaml.FullLoader resolves ``!!python/name:os.system`` to the live
+        function; safe loading must reject the tag at parse time instead.
+        """
+        tagged = dict(sample_config_dict)
+        tagged["description"] = "PLACEHOLDER_FOR_TAG"
+        path = _write_yaml(tmp_path, tagged)
+        text = path.read_text(encoding="utf-8")
+        path.write_text(
+            text.replace("PLACEHOLDER_FOR_TAG", "!!python/name:os.system"),
+            encoding="utf-8",
+        )
+
+        with pytest.raises(yaml.YAMLError):
+            SUEWSConfig.from_yaml(str(path))
+
+    def test_unknown_top_level_key_raises(self, sample_config_dict):
+        """Unknown top-level keys must raise, not be silently retained."""
+        drifted = dict(sample_config_dict)
+        drifted["stray_key"] = 1
+
+        with pytest.raises(ValueError, match="stray_key"):
+            SUEWSConfig.from_dict(drifted)
+
+    def test_internal_bookkeeping_keys_accepted(self, sample_config_dict, tmp_path):
+        """The private _yaml_* keys must keep working through from_yaml."""
+        path = _write_yaml(tmp_path, sample_config_dict)
+        config = SUEWSConfig.from_yaml(str(path))
+        assert getattr(config, "_yaml_path", None) == str(path)
+
+    def test_assignment_is_validated(self, sample_config_dict):
+        """Direct attribute assignment must be validated, not stored raw."""
+        config = SUEWSConfig.from_dict(sample_config_dict)
+
+        with pytest.raises(ValueError):
+            config.model.physics.net_radiation = "not_a_method"
+
+        with pytest.raises(ValueError):
+            config.sites = "not a list"
+
+        # Valid assignments still coerce and succeed
+        config.model.control.tstep = 600
+        tstep = config.model.control.tstep
+        assert int(tstep.value if hasattr(tstep, "value") else tstep) == 600
+
+    def test_from_dict_runs_completeness_checks_like_yaml(
+        self, sample_config_dict, tmp_path
+    ):
+        """Dict input must hit the same site-completeness checks as YAML.
+
+        gh#1530 review follow-up: the completeness validator was gated on
+        ``_yaml_path``, so the in-memory dict path silently skipped checks
+        that the same data loaded from a file would fail.
+        """
+        from copy import deepcopy
+
+        broken = deepcopy(sample_config_dict)
+        broken["sites"][0]["properties"]["land_cover"]["grass"].pop("lai")
+
+        path = _write_yaml(tmp_path, broken)
+        with pytest.raises(ValueError):
+            SUEWSConfig.from_yaml(str(path))
+
+        with pytest.raises(ValueError):
+            SUEWSConfig.from_dict(broken)
+
+# ===========================================================================
+# to_yaml round-trips -- formerly test_to_yaml_roundtrip.py
+# ===========================================================================
 
 @pytest.fixture
 def standard_data():
